requirements.txt
pandas==2.2.3
numpy==2.2.6
openpyxl==3.1.5
pyyaml==6.0.2
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.2
hydra-core==1.3.2
omegaconf==2.3.0
dvc==3.59.2
dvc-s3==3.2.0
mlflow-skinny==2.22.0
wandb==0.19.11
fastapi==0.110.0
uvicorn==0.29.0
requests==2.32.3
